backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asgi-lifespan==2.1.0
asyncpg==0.31.0
attrs==25.4.0
cachetools==6.2.4
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography>=41,<43
cycler==0.12.1
fastapi==0.125.0
fonttools==4.61.1
frozenlist==1.8.0
google-auth==2.45.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
joblib==1.5.3
kiwisolver==1.4.9
matplotlib==3.10.8
msgpack==1.1.2
multidict==6.7.0
numpy==2.3.5
packaging==25.0
pandas==2.3.3
pillow==12.0.0
propcache==0.4.1
protobuf==6.33.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
Pympler==1.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
requests==2.32.5
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.16.3
seaborn==0.13.2
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
statbotics==3.0.0
threadpoolctl==3.6.0
tqdm==4.67.1
ttkbootstrap==1.19.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.38.0
yarl==1.22.0
pywebpush==1.14.0

